[SPARK-38006][SQL] Clean up duplicated planner logic for window operators
### What changes were proposed in this pull request?

Both `WindowExec.scala` and `WindowInPandasExec.scala` have some duplicated logic with regarded to query planning (`output`, `requiredChildDistribution/Ordering`, `outputPartitioning/Ordering`). We can move these logic into their common parent `WindowExecBase` to reduce duplication.

### Why are the changes needed?

Clean up existing code for better readability.

### Does this PR introduce _any_ user-facing change?

No.

### How was this patch tested?

Existing test as this is just refactoring.

Closes #35305 from c21/window-fix.

Authored-by: Cheng Su <chengsu@fb.com>
